requirements.txt
jmcomic~=2.5.32
uvicorn~=0.34.0
fastapi~=0.115.12
starlette~=0.46.1
pydantic~=2.10.6
pillow~=11.1.0